extrect some logic to a service helper struct
